fixed various bugs

+++ src/interface_adapter/LoggedIn/AddExercise/AddExerciseState.java
@@ -3,7 +3,7 @@ package interface_adapter.LoggedIn.AddExercise;
 
 public class AddExerciseState {
     private String exerciseName = "";
-    private float exerciseLength = 0;
+    private double exerciseLength = 0;
     private String exerciseError = null;
 
 
@@ -23,11 +23,11 @@ public class AddExerciseState {
         this.exerciseName = exerciseName;
     }
 
-    public float getExerciseLength() {
+    public double getExerciseLength() {
         return exerciseLength;
     }
 
-    public void setExerciseLength(float exerciseLength) {
+    public void setExerciseLength(double exerciseLength) {
         this.exerciseLength = exerciseLength;
     }
